Finished cleaning up unit tests for the migrated toast code, and updated VueJS' version to be able to use useTempalteRef
