+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.5'
-
 services:
   reverse-proxy:
     image: traefik
