+++ frontend/Dockerfile
@@ -1,5 +1,5 @@
 # Local development stage
-FROM node:20.8-alpine3.18 as development-stage
+FROM node:20.8-alpine3.18 AS development-stage
 WORKDIR /app
 COPY package.json /app/
 COPY yarn.lock /app/
@@ -10,7 +10,7 @@ EXPOSE 3000
 ENTRYPOINT ["yarn", "dev:host"]
 
 # Production Build stage
-FROM node:20.8-alpine3.18 as production-build
+FROM node:20.8-alpine3.18 AS production-build
 WORKDIR /app
 COPY ./src /app/src/
 COPY package.json /app/
@@ -21,9 +21,9 @@ ARG VERSION_BUILD_TAG=latest
 
 ENV VITE_ROSALUTION_VERSION=$VERSION_BUILD_TAG
 
-RUN yarn install --frozen-lockfile && yarn build --base=/rosalution/
+RUN yarn install --frozen-lockfile && yarn build --bASe=/rosalution/
 
-FROM nginx:1.25.2-alpine3.18 as production-stage
+FROM nginx:1.25.2-alpine3.18 AS production-stage
 
 COPY etc/default.conf /etc/nginx/conf.d/
 COPY --from=production-build /app/dist/ /usr/share/nginx/html/
