+++ frontend/package.json
@@ -21,7 +21,7 @@
     "@fortawesome/free-solid-svg-icons": "6.4.2",
     "@fortawesome/vue-fontawesome": "3.0.3",
     "@rollup/plugin-strip": "3.0.4",
-    "vue": "3.3.4",
+    "vue": "3.5.12",
     "vue-router": "4.2.5"
   },
   "devDependencies": {
