+++ frontend/test/components/Dialogs/RosalutionToast.spec.js
@@ -4,25 +4,29 @@ import sinon from 'sinon';
 
 import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
 
-import Toast from '@/components/Dialogs/Toast.vue';
-import toast from '@/toast.js';
+import RosalutionToast from '@/components/Dialogs/RosalutionToast.vue';
 
-describe('Dialog.vue', () => {
+describe('Toast.vue', () => {
   let wrapper;
   let sandbox;
+  let toast;
 
   beforeAll(() => {
     sandbox = sinon.createSandbox();
   });
 
   beforeEach(() => {
-    wrapper = shallowMount(Toast, {
+    wrapper = shallowMount(RosalutionToast, {
       global: {
         components: {
           'font-awesome-icon': FontAwesomeIcon,
         },
+        stubs: {
+          transition: false,
+        },
       },
     });
+    toast = wrapper.vm;
   });
 
   afterEach(() => {
@@ -63,13 +67,28 @@ describe('Dialog.vue', () => {
   it('should allow the user to close', async () => {
     toast.info('hello world');
     await wrapper.vm.$nextTick();
+    const toastContainer = wrapper.find('[data-test=toast-container]');
+    expect(toastContainer.exists()).to.be.true;
 
     const closeIconElement = wrapper.get('[data-test=toast-close-button]');
-    closeIconElement.trigger('click');
+    await closeIconElement.trigger('click');
+
+    const closedContainer = wrapper.find('[data-test=toast-container]');
+    expect(closedContainer.exists()).to.be.false;
+  });
+
+  it('displays different and multiple prompts to the toast', async () => {
+    toast.info('hello world first time');
     await wrapper.vm.$nextTick();
 
-    const toastContainer = wrapper.find('[data-test=toast-container]');
-    expect(toastContainer.exists()).to.be.false;
+    expect(wrapper.html()).to.include('hello world first time');
+
+    toast.error('hello world error');
+    await wrapper.vm.$nextTick();
+    expect(wrapper.html()).to.include('hello world error');
+
+    toast.cancel();
+    await wrapper.vm.$nextTick();
   });
 });
